+++ frontend/src/App.js
@@ -1,9 +1,9 @@
 //App.js는 앱 안에서 페이지를 구성하고 라우팅을 결정하는 역할
 import React from 'react';
 import {Route, Routes} from 'react-router-dom';
-import ChatRoom from './chatroom';
+import ChatRoom from './pages/chat/chatroom';
 import Home from './Home';
-import ChatListScreen from "./chatList/chatListScreen";
+import ChatListScreen from "./pages/chat/chatListScreen";
 import Login from './pages/auth/Login';
 import Register from "./pages/auth/Register";
 import FindId from "./pages/auth/FindId";
